Clean up  code
