+++ lib/Screens/Home/chat_screens/chat_screens.dart
@@ -5,8 +5,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:helpdesk_2/Screens/Home/chat_screens/widgets/cached_image.dart';
+import 'package:helpdesk_2/core/common/image_utils.dart';
 import 'package:helpdesk_2/core/enum/view_state.dart';
-import 'package:helpdesk_2/core/utils/image_utils.dart';
 import 'package:helpdesk_2/core/widget/appbar.dart';
 import 'package:helpdesk_2/data/db/models/message.dart';
 import 'package:helpdesk_2/main.dart';
