+++ lib/core/common/image_utils.dart
@@ -1,7 +1,5 @@
 import 'dart:io';
 import 'dart:math';
-
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:image/image.dart' as Im;
