+++ lib/core/common/validator.dart
@@ -1,0 +1,39 @@
+
+
+class NameValidator {
+  static String validate(String value) {
+    if (value.isEmpty) {
+      return "Name can't be empty";
+    }
+    if (value.length < 2) {
+      return "Name must be at least 2 characters long";
+    }
+
+    if (value.length > 50) {
+      return "Name can't be more than 50 characters";
+    }
+    return null;
+  }
+}
+
+class PasswordValidator {
+  static String validate(String value) {
+    if (value.isEmpty) {
+      return "Password can't be empty";
+    }
+    return null;
+  }
+}
+
+class EmailValidator {
+  static String validate(String value) {
+    if (value.isEmpty) {
+      return "Email can't be empty";
+    }
+    // if(!value.contains("@")){
+    //   return "Please enter a valid email!";
+    // }
+
+    return null;
+  }
+}
