+++ lib/provider/user_provider.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/widgets.dart';
 import 'package:helpdesk_2/data/db/models/helper.dart';
-import 'package:helpdesk_2/data/service/auth_services.dart';
+import 'package:helpdesk_2/data/service/auth_service.dart';
 
 class UserProvider with ChangeNotifier {
   Helper _helper;
-  AuthServices _authServices = AuthServices();
+  AuthService _authServices = AuthService();
 
   Helper get getHelper => _helper;
